feat: move files around + finish typing
