+++ hatchet_sdk/worker/worker.py
@@ -113,11 +113,10 @@ class Worker:
             logger.error(e)
             sys.exit(1)
 
-    def register_workflow(self, workflow: TWorkflow) -> None:
-        ## Hack for typing
-        assert isinstance(workflow, WorkflowInterface)
-
+    def register_workflow(self, workflow) -> None:
         namespace = self.client.config.namespace
+
+        print(f"registering workflow: {workflow}", workflow.steps)
 
         try:
             self.client.admin.put_workflow(
